moved Geo conversions into a separate source file
