+++ Assets/ACViewer-SDK/AugmentedCity/Scripts/GeoMath.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e08bae65578db4c71ac6fd413a1cc49e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
